+++ module_groups/elastic_ci_stack/builders_stack.tf
@@ -1,12 +1,13 @@
 module "builders_stack" {
   source = "/deployer/modules/buildkite_elastic_stack"
 
+  agents_per_instance = "${var.builder_agents_per_instance}"
   bootstrap_script_url = "${var.builder_bootstrap_script_url}"
   buildkite_agent_token = "${var.buildkite_agent_token}"
   buildkite_queue = "${var.buildkite_queue_prefix}-builders"
-  instance_type = "c5.large"
+  instance_type = "${var.builder_instance_type}"
   key_name = "${aws_key_pair.key_pair.key_name}"
-  max_size = "1"
+  max_size = "${var.builder_max_size}"
   min_size = "1"
   secrets_bucket = "${var.secrets_bucket}"
   stack_ami_version = "${var.stack_ami_version}"